Comments of doxygen
On update

+++ cpp/samchon/library/Event.cpp
@@ -2,7 +2,7 @@
 #include <samchon/library/EventDispatcher.hpp>
 using namespace samchon::library;
 
-Event::Event(EventDispatcher *source, long type)
+Event::Event(EventDispatcher *source, int type)
 {
 	this->source = source;
 	this->type = type;
@@ -12,7 +12,7 @@ auto Event::getSource() const -> EventDispatcher*
 {
 	return source;
 }
-auto Event::getType() const -> long
+auto Event::getType() const -> int
 {
 	return type;
 }
